refactor(skills): simplify writeSkills after over-engineering review
- drop the write-order sort in writeFiles; filesystem write order is
  unobservable and the final tree is identical
- detect source name collisions inside the bundle loop instead of a
  separate dedupe pass
- inline the single-use PreparedSkill interface
- fold renderDistributionMetadata into render.ts and delete the
  single-function metadata.ts module
