+++ packages/skills/src/render.ts
@@ -5,7 +5,32 @@
 import { posix } from "node:path";
 
 import { normalizeInstructionList } from "./annotations.ts";
-import type { ManifestArg, ManifestFlag, ManifestNode, RenderedFile, SkillMeta } from "./types.ts";
+import type {
+	ManifestArg,
+	ManifestFlag,
+	ManifestNode,
+	RenderedFile,
+	SkillKind,
+	SkillMeta,
+} from "./types.ts";
+import { CRUST_MANIFEST } from "./version.ts";
+
+/** Renders the ownership and version metadata shipped with a skill. */
+export function renderDistributionMetadata(meta: SkillMeta, kind: SkillKind): RenderedFile {
+	return {
+		path: CRUST_MANIFEST,
+		content: `${JSON.stringify(
+			{
+				name: meta.name,
+				description: meta.description,
+				version: meta.version,
+				kind,
+			},
+			null,
+			"\t",
+		)}\n`,
+	};
+}
 
 // ────────────────────────────────────────────────────────────────────────────
 // Text escaping helpers
